Move SqlTaskStatus and SqlTaskStausTest from msq module to sql module.
This is required for the scheduled batch supervisor.
