+++ sql/src/main/java/org/apache/druid/sql/http/SqlTaskStatus.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.druid.msq.sql;
+package org.apache.druid.sql.http;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -25,13 +25,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Preconditions;
 import org.apache.druid.error.ErrorResponse;
 import org.apache.druid.indexer.TaskState;
-import org.apache.druid.msq.sql.resources.SqlTaskResource;
 
 import javax.annotation.Nullable;
 import java.util.Objects;
 
 /**
- * Response object for {@link SqlTaskResource#doPost}.
+ * Response object for org.apache.druid.msq.sql.resources.SqlTaskResource#doPost.
  */
 public class SqlTaskStatus
 {
